+++ packages/web3-eth/test/unit/rpc_method_wrappers/get_code.test.ts
@@ -67,7 +67,7 @@ describe('getCode', () => {
 	it.each(testData)(
 		`should format mockRpcResponse using provided return format\nTitle: %s\nInput parameters: %s\n`,
 		async (_, inputParameters) => {
-			const expectedReturnFormat = { number: FMT_NUMBER.STR, bytes: FMT_BYTES.BUFFER };
+			const expectedReturnFormat = { number: FMT_NUMBER.STR, bytes: FMT_BYTES.UINT8ARRAY };
 			const expectedFormattedResult = format(
 				{ eth: 'bytes' },
 				mockRpcResponse,