+++ packages/web3/test/e2e/sepolia/call.test.ts
@@ -14,7 +14,7 @@ GNU Lesser General Public License for more details.
 You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
-import { bytesToBuffer, hexToBytes } from 'web3-utils';
+import { hexToBytes } from 'web3-utils';
 
 import Web3, { FMT_BYTES, FMT_NUMBER } from '../../../src';
 import {
@@ -63,16 +63,6 @@ describe(`${getSystemTestBackend()} tests - call`, () => {
 					'0x0000000000000000000000000000000000000000000000000000000000000000',
 				);
 				break;
-			case 'BYTES_BUFFER':
-				// eslint-disable-next-line jest/no-conditional-expect
-				expect(result).toStrictEqual(
-					bytesToBuffer(
-						hexToBytes(
-							'0x0000000000000000000000000000000000000000000000000000000000000000',
-						),
-					),
-				);
-				break;
 			case 'BYTES_UINT8ARRAY':
 				// eslint-disable-next-line jest/no-conditional-expect
 				expect(result).toStrictEqual(
@@ -114,10 +104,6 @@ describe(`${getSystemTestBackend()} tests - call`, () => {
 			case 'BYTES_HEX':
 				// eslint-disable-next-line jest/no-conditional-expect
 				expect(result).toBe(expectedResult);
-				break;
-			case 'BYTES_BUFFER':
-				// eslint-disable-next-line jest/no-conditional-expect
-				expect(result).toStrictEqual(bytesToBuffer(hexToBytes(expectedResult)));
 				break;
 			case 'BYTES_UINT8ARRAY':
 				// eslint-disable-next-line jest/no-conditional-expect