+++ templates/cypress/webpack.config.js
@@ -29,9 +29,6 @@ module.exports = {
 	},
 	plugins: [
 		new webpack.ProvidePlugin({
-			Buffer: ['buffer', 'Buffer'],
-		}),
-		new webpack.ProvidePlugin({
 			process: 'process/browser',
 		}),
 	],