+++ packages/web3-types/CHANGELOG.md
@@ -102,3 +102,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   Added `filters` param to the `Filter` type (#6010)
+
+### Changed
+
+-   Replaced Buffer for Uint8Array (#6004)
+-   types `FMT_BYTES.BUFFER`, `Bytes` and `FormatType` and encryption option types for `salt` and `iv` has replaced support for `Buffer` for `Uint8Array` (#6004)